requirements.txt
asgiref==3.4.1
APScheduler==3.7.0
charset-normalizer==2.0.4
channels==3.0.4
daphne==3.0.2
django==3.1.7
django-allauth==0.44.0
django-apscheduler==0.6.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
djangorestframework-jwt==1.11.0
requests==2.26.0
pycparser==2.20
PyJWT==1.7.1
mysqlclient==2.0.3
PyMySQL==1.0.2
pyOpenSSL==20.0.1
urllib3==1.26.6
chardet==4.0.0
sqlparse==0.4.1
six==1.16.0
idna==3.2
h11==0.12.0
pytz==2020.4
pillow==8.4.0